+++ controller/api/main_controller.go
@@ -1,4 +1,4 @@
-package controller
+package api
 
 import(
 	"github.com/kataras/iris/v12/mvc"
